+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -800,7 +800,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             }
 
             for (AnnotationMirror poly : polys) {
-                if (type.hasAnnotationRelaxed(poly)) {
+                if (type.hasAnnotation(poly)) {
                     return Collections.singletonList(
                             DiagMessage.error("invalid.polymorphic.qualifier.use", poly));
                 }
@@ -2210,7 +2210,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
 
             CFAbstractStore<?, ?> store = atypeFactory.getStoreBefore(tree);
 
-            Set<AnnotationMirror> annos =
+            AnnotationMirrorSet annos =
                     atypeFactory.getAnnotatedTypeBefore(exprJe, tree).getAnnotations();
 
             AnnotationMirror inferredAnno =
@@ -3149,7 +3149,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
         List<AnnotationTree> result = new ArrayList<>(1);
         for (AnnotationTree at : annoTrees) {
             AnnotationMirror anno = TreeUtils.annotationFromAnnotationTree(at);
-            if (!AnnotationUtils.isDeclarationAnnotation(anno)
+            if (AnnotationUtils.isTypeUseAnnotation(anno)
                     && atypeFactory.isSupportedQualifier(anno)) {
                 result.add(at);
             }