+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
-Version 3.42.0-eisop2 (January ?, 2024)
----------------------------------------
+Version 3.42.0-eisop5 (July ?, 2024)
+------------------------------------
 
 **User-visible changes:**
 
@@ -13,6 +13,70 @@ it is an upcast, downcast or incomparable cast. Based on that, further determine
 is statically verifiable or not.
 
 **Closed issues:**
+
+
+Version 3.42.0-eisop4 (July 12, 2024)
+-------------------------------------
+
+**Implementation details:**
+
+New method `GenericAnnotatedTypeFactory#addComputedTypeAnnotationsWithoutFlow(Tree, AnnotatedTypeMirror)`
+that sets `useFlow` to `false` before calling `addComputedTypeAnnotations`. Subclasses should override
+method `GenericAnnotatedTypeFactory#addComputedTypeAnnotations(Tree, AnnotatedTypeMirror)` instead.
+Deprecated the `GenericAnnotatedTypeFactory#addComputedTypeAnnotations(Tree, AnnotatedTypeMirror, boolean)`
+overload.
+
+Changed the return type of `AnnotatedTypeFactory#getEnumConstructorQualifiers` from `Set<AnnotationMirror>`
+to `AnnotationMirrorSet`.
+
+framework-test:
+- Improvements to more consistently handle tests that do not use `-Anomsgtext`.
+- Added new class `DetailedTestDiagnostic` to directly represent test diagnostics when
+  `-Adetailedmsgtext` is used.
+
+**Closed issues:**
+
+eisop#742, eisop#777, eisop#795, typetools#6704.
+
+
+Version 3.42.0-eisop3 (March 1, 2024)
+-------------------------------------
+
+**User-visible changes:**
+
+Performance improvements in the Nullness Checker.
+
+**Implementation details:**
+
+Support separate defaults for wildcard and type variable upper bounds.
+Add support for defaults for type variable uses.
+See changes in `TypeUseLocation`, `QualiferDefaults`, and `QualifierHierarchy`,
+as well as the new `ParametricTypeVariableUseQualifier` meta-annotation.
+
+Refactored the `TypeInformationPresenter` into several classes in the new
+`org.checkerframework.framework.util.visualize` package.
+
+**Closed issues:**
+
+eisop#703, typetools#6433, typetools#6438.
+
+
+Version 3.42.0-eisop2 (January 9, 2024)
+---------------------------------------
+
+**Implementation details:**
+
+Moved `ErrorTypeKindException` from `org.checkerframework.framework.util.element.ElementAnnotationUtil` to
+`org.checkerframework.framework.type.AnnotatedTypeMirror`. Properly raise these errors in more cases.
+
+Deprecated `AnnotationUtils#isDeclarationAnnotation` and added the clearer `AnnotationUtils#isTypeUseAnnotation`.
+
+Removed the dependency on the classgraph library, which added over 500kB to `checker.jar`.
+It is easy to add the dependency for debugging.
+
+**Closed issues:**
+
+eisop#666, eisop#673.
 
 
 Version 3.42.0-eisop1 (January 2, 2024)